+++ core/action-queue/tests/auto_queue.rs
@@ -1,14 +1,17 @@
 #![allow(clippy::ignored_unit_patterns)]
 mod common;
 
+use std::time::Duration;
+
 use crate::common::DefaultError;
 use common::new_queue_typed;
 use proton_action_queue::action::{
     Action, ActionId, DefaultVersionConverter, Handler, Type, WriterGuard, WriterGuardError,
 };
-use proton_action_queue::queue::QueuedActionState;
+use proton_action_queue::queue::{BroadcastMessage, QueuedActionState};
 use serde::{Deserialize, Serialize};
 use stash::stash::Bond;
+use tokio::time::sleep;
 
 #[tokio::test]
 async fn auto_queued_on_network_failure() {
@@ -19,6 +22,121 @@ async fn auto_queued_on_network_failure() {
     let output = queue.new_executor().execute_one().await.unwrap().unwrap();
 
     assert!(matches!(output, QueuedActionState::Queued(_)),);
+}
+
+#[tokio::test]
+async fn auto_queued_on_pause() {
+    let queue = new_queue_typed::<SuccessAction>().await;
+    let mut broadcast = queue.new_broadcast_receiver();
+    let auto_executor = queue.new_executor().into_auto_executor();
+
+    auto_executor.pause();
+    queue.queue_action(SuccessAction {}).await.unwrap();
+
+    sleep(Duration::from_secs(1)).await;
+    assert!(broadcast.is_empty());
+
+    auto_executor.unpause();
+
+    let output = broadcast.recv().await.unwrap();
+
+    assert!(matches!(output, BroadcastMessage::Success(_)));
+    assert!(broadcast.is_empty());
+
+    auto_executor.terminate();
+    sleep(Duration::from_millis(100)).await;
+
+    assert_eq!(queue.queued_actions_count().await.unwrap(), 0);
+}
+
+#[tokio::test]
+async fn auto_queued_on_multiple_unpause() {
+    let queue = new_queue_typed::<SuccessAction>().await;
+    let mut broadcast = queue.new_broadcast_receiver();
+
+    queue.queue_action(SuccessAction {}).await.unwrap();
+
+    let auto_executor = queue.new_executor().into_auto_executor();
+
+    // Calling unpause should have no effect as auto executors starts unpaused.
+    auto_executor.unpause();
+    auto_executor.unpause();
+    auto_executor.unpause();
+
+    let output = broadcast.recv().await.unwrap();
+
+    assert!(matches!(output, BroadcastMessage::Success(_)));
+    assert!(broadcast.is_empty());
+
+    auto_executor.terminate();
+    sleep(Duration::from_millis(100)).await;
+
+    assert_eq!(queue.queued_actions_count().await.unwrap(), 0);
+}
+
+#[tokio::test]
+async fn auto_queued_on_multiple_pause() {
+    let queue = new_queue_typed::<SuccessAction>().await;
+    let mut broadcast = queue.new_broadcast_receiver();
+    let auto_executor = queue.new_executor().into_auto_executor();
+
+    // Calling pause multiple times should still end up in paused state.
+    auto_executor.pause();
+    auto_executor.pause();
+    auto_executor.pause();
+    queue.queue_action(SuccessAction {}).await.unwrap();
+
+    sleep(Duration::from_secs(1)).await;
+    assert!(broadcast.is_empty());
+
+    // Calling unpause multiple times should always end up in upaused state.
+    auto_executor.unpause();
+    auto_executor.unpause();
+    auto_executor.unpause();
+
+    let output = broadcast.recv().await.unwrap();
+
+    assert!(matches!(output, BroadcastMessage::Success(_)));
+    assert!(broadcast.is_empty());
+
+    auto_executor.terminate();
+    sleep(Duration::from_millis(100)).await;
+
+    assert_eq!(queue.queued_actions_count().await.unwrap(), 0);
+}
+
+#[tokio::test]
+async fn auto_queued_on_pause_and_partially_manual_execution() {
+    let queue = new_queue_typed::<SuccessAction>().await;
+    let mut broadcast = queue.new_broadcast_receiver();
+    let auto_executor = queue.new_executor().into_auto_executor();
+
+    auto_executor.pause();
+    queue.queue_action(SuccessAction {}).await.unwrap();
+    queue.queue_action(SuccessAction {}).await.unwrap();
+
+    sleep(Duration::from_secs(1)).await;
+    assert!(broadcast.is_empty());
+
+    let output = queue.new_executor().execute_one().await.unwrap().unwrap();
+
+    assert!(matches!(output, QueuedActionState::Executed(_)),);
+
+    let output = broadcast.recv().await.unwrap();
+
+    assert!(matches!(output, BroadcastMessage::Success(_)));
+
+    auto_executor.unpause();
+
+    let output = broadcast.recv().await.unwrap();
+
+    assert!(matches!(output, BroadcastMessage::Success(_)));
+    assert!(broadcast.is_empty());
+
+    auto_executor.terminate();
+    sleep(Duration::from_millis(100)).await;
+
+    assert_eq!(queue.queued_actions_count().await.unwrap(), 0);
 }
 
 #[tokio::test]
@@ -44,6 +162,60 @@ async fn execute_all_does_not_loop_forever_on_network_failure() {
     let _ = queue.queue_action(ErrorAction {}).await.unwrap();
 
     queue.new_executor().execute_all().await.unwrap();
+}
+
+#[derive(Serialize, Deserialize)]
+struct SuccessAction {}
+
+impl Action for SuccessAction {
+    const TYPE: Type = Type("success");
+    const VERSION: u32 = 1;
+    type VersionConverter = DefaultVersionConverter<Self>;
+    type Handler = SuccessActionHandler;
+    type RemoteOutput = ();
+    type LocalOutput = ();
+    type Error = DefaultError;
+    type Context = ();
+}
+
+#[derive(Default)]
+struct SuccessActionHandler {}
+
+impl Handler for SuccessActionHandler {
+    type Action = SuccessAction;
+
+    type Context = ();
+
+    async fn apply_local(
+        &self,
+        _: ActionId,
+        _: &Self::Context,
+        _: &mut Self::Action,
+        _: &Bond<'_>,
+    ) -> Result<(), <Self::Action as Action>::Error> {
+        // Nothing to do
+        Ok(())
+    }
+
+    async fn revert_local(
+        &self,
+        _: ActionId,
+        _: &Self::Context,
+        _: &mut Self::Action,
+        _: &Bond<'_>,
+    ) -> Result<(), <Self::Action as Action>::Error> {
+        panic!("should not be called");
+    }
+
+    async fn apply_remote(
+        &self,
+        _: ActionId,
+        _: &Self::Context,
+        _: &mut Self::Action,
+        _: WriterGuard<'_>,
+    ) -> Result<<Self::Action as Action>::RemoteOutput, <Self::Action as Action>::Error> {
+        Ok(())
+    }
 }
 
 #[derive(Serialize, Deserialize)]