+++ mail/mail-common/src/user_context.rs
@@ -130,6 +130,18 @@ impl MailUserContext {
         self.user_context.queue()
     }
 
+    /// Pause all action queue executors.
+    pub fn pause_queue_executors(&self) {
+        self.user_context.queue().queue_executor.pause();
+        self.send_queue_executors.pause();
+    }
+
+    /// Unpause all action queue executors.
+    pub fn unpause_queue_executors(&self) {
+        self.user_context.queue().queue_executor.unpause();
+        self.send_queue_executors.unpause();
+    }
+
     /// Terminate all action queue executors.
     pub fn terminate_queue_executors(&self) {
         self.user_context.queue().queue_executor.terminate();